requirements.txt.txt
discord.py==2.6.0
python-dotenv==1.0.0